+++ src/Console/ExitCode.php
@@ -7,6 +7,7 @@ class ExitCode
     // SugarCli errors
     const EXIT_STATUS_MODIFICATIONS = 2;
     const EXIT_FORMAT_ERROR = 3;
+    const EXIT_INVENTORY_ERROR = 4;
 
     // Sugar installer errors
     const EXIT_NOT_EXTRACTED = 11;